feat: docker compose and languages support

+++ CodePlayground.Core/Models/CodeExecutionResult.cs
@@ -4,5 +4,5 @@ public class CodeExecutionResult
 {
     public string? StdOut { get; set; }
     public string? StdErr { get; set; }
-    public int ExitCode { get; set; }
+    public long ExitCode { get; set; }
 }
